Made Closed Loop Calibration on IOCouplers only.
It turns out that Open Loop Calibration for the Multipliers is still good.  In this version, in the Closed Loop Calibration, we simply calibrate for an "average" input/output coupler, and then again allowing all of the individual input/output coupler values to vary.  The alignment with the closed loop data run appears to be very good.

Also, some minor linting.